Rename "parser" module to "entity_parser"
