+++ snips_nlu/entity_parser/__init__.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 from __future__ import unicode_literals
 
-from snips_nlu.parser.builtin_entity_parser import (
+from snips_nlu.entity_parser.builtin_entity_parser import (
     BuiltinEntityParser, get_builtin_entity_parser)
-from snips_nlu.parser.custom_entity_parser import (
+from snips_nlu.entity_parser.custom_entity_parser import (
     CustomEntityParser, get_custom_entity_parser)
