+++ snips_nlu/tests/test_builtin_entity_parser.py
